committing files

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,9 +8,9 @@
 */
 int main(void)
 {
-	long long int a = 1;
-	long long int b = 2;
-	long long int s = 0;
+	unsigned int a = 1;
+	unsigned int b = 2;
+	unsigned int s = 0;
 	int i;
 
 	for (i = 0 ; i <= 49 ; i++)
@@ -18,7 +18,7 @@ int main(void)
 		s = a + b;
 		a = b;
 		b = s;
-		printf("%lld", s);
+		printf("%lu", s);
 		if (i != 49)
 		{
 			putchar(',');
